Add basic tests
