--- src/main/java/edu/itu/cavabunga/entities/Parameter.java
@@ -1,4 +1,0 @@
-package edu.itu.cavabunga.entities;
-
-public class Parameter {
-}
