--- src/main/java/edu/itu/cavabunga/entities/Property.java
@@ -1,4 +1,0 @@
-package edu.itu.cavabunga.entities;
-
-public class Property {
-}
