+++ src/main/java/edu/itu/cavabunga/entity/Parameter.java
@@ -1,0 +1,4 @@
+package edu.itu.cavabunga.entity;
+
+public class Parameter {
+}
