+++ src/main/java/edu/itu/cavabunga/entity/Property.java
@@ -1,0 +1,4 @@
+package edu.itu.cavabunga.entity;
+
+public class Property {
+}
