--- src/main/java/edu/itu/cavabunga/repositories/CalendarRepository.java
@@ -1,7 +1,0 @@
-package edu.itu.cavabunga.repositories;
-
-import edu.itu.cavabunga.entities.Calendar;
-import org.springframework.data.repository.CrudRepository;
-
-public interface CalendarRepository extends CrudRepository<Calendar, Integer> {
-}
