+++ src/main/java/edu/itu/cavabunga/repository/CalendarRepository.java
@@ -1,0 +1,7 @@
+package edu.itu.cavabunga.repository;
+
+import edu.itu.cavabunga.entity.Calendar;
+import org.springframework.data.repository.CrudRepository;
+
+public interface CalendarRepository extends CrudRepository<Calendar, Long> {
+}
